+++ packages/kit/src/constants/localStorage.ts
@@ -11,6 +11,7 @@ const WAAS_APPLE_ID_TOKEN = 'waasAppleIdToken'
 const WAAS_SESSION_HASH = 'waasSessionHash'
 const WAAS_EMAIL_ID_TOKEN = 'waasEmailIdToken'
 const WAAS_ACTIVE_LOGIN_TYPE = 'waasActiveLoginType'
+const WAAS_SIGN_IN_EMAIL = 'waasSignInEmail'
 
 // TODO: remove all of this.. we should not be storing these in local storage
 export enum LocalStorageKey {
@@ -25,5 +26,6 @@ export enum LocalStorageKey {
   WaasAppleIdToken = `${PREFIX}.${WAAS_APPLE_ID_TOKEN}`,
   WaasSessionHash = `${PREFIX}.${WAAS_SESSION_HASH}`,
   WaasActiveLoginType = `${PREFIX}.${WAAS_ACTIVE_LOGIN_TYPE}`,
-  WaasEmailIdToken = `${PREFIX}.${WAAS_EMAIL_ID_TOKEN}`
+  WaasEmailIdToken = `${PREFIX}.${WAAS_EMAIL_ID_TOKEN}`,
+  WaasSignInEmail = `${PREFIX}.${WAAS_SIGN_IN_EMAIL}`,
 }